更新Payload第一个字节为status 并增加license

+++ pc-sdk/csharp/NexLinkNative.cs
@@ -37,7 +37,7 @@ namespace NexLink
         [DllImport(DLL, CallingConvention = CallingConvention.Cdecl)]
         public static extern int nexlink_send_async(
             IntPtr handle,
-            byte cmd,
+            ushort cmd,
             byte[] payload,
             ushort length);
